Rename MavenCentralDownloader to MavenRepository, refactored for less redundant code, more unit testable
